+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * added ```spiders.py``` to enable infrastructure spider discovery
 * added ```get-clf-version.sh``` the encapsulate the pattern of parsing ```setup.py```
   to extract the cloudfeaster version
+* added runtime docker image for running spiders
 
 ### Changed